+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 
 - `datadogexporter`: add error checks for datadog exporter (#9964)
 - `groupbyattrsprocessor`: copied aggregationtemporality when grouping metrics. (#9088)
+- `mongodbreceiver`: Fix issue where receiver startup could hang (#10111)
 
 ## v0.51.0